Re-structure the project files and directories
