+++ src/App.js
@@ -1,11 +1,13 @@
-import Calculator from './components/Calculator';
-import Home from './components/Home';
-import Quote from './components/Quote';
-import Header from './components/Header';
-import Footer from './components/Footer';
-
 // React Router
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+
+// App pages
+import Calculator from './pages/Calculator';
+import Home from './pages/Home';
+import Quote from './pages/Quote';
+// Components
+import Header from './components/Header';
+import Footer from './components/Footer';
 
 // APP
 const App = () => {
