+++ src/components/Header.jsx
@@ -1,4 +1,4 @@
-import "./Header.css";
+import "../styles/Header.css";
 
 const Header = () => {
   const FooterBackground = {
