+++ src/pages/Calculator.jsx
@@ -1,31 +1,11 @@
 import { useState } from 'react';
-import PropTypes from 'prop-types';
+// Logic
 import calculate from "../logic/calculate";
+// Calculator components
+import { Screen, Number, Oprand } from '../components/CalculatorParts';
 
 // Stylesheets
-import "./Calculator.css"
-
-// Sub Components => {Screen, oprands, Numbers}
-const Screen = ({ result }) => <span className="screen">{result}</span>;
-Screen.propTypes = { result: PropTypes.string.isRequired };
-
-const Oprand = ({ data, className, onclick}) => (
-  <button type="button" className={className} onClick={onclick}>
-    {data}
-  </button>
-);
-Oprand.propTypes = { data: PropTypes.string.isRequired };
-Oprand.propTypes = { className: PropTypes.string };
-Oprand.propTypes = { onclick: PropTypes.func.isRequired };
-
-const Number = ({ value, className , onclick}) => (
-  <button type="button" className={className} onClick={onclick}>
-    {value}
-  </button>
-);
-Number.propTypes = { value: PropTypes.string.isRequired };
-Number.propTypes = { className: PropTypes.string};
-Number.propTypes = { onclick: PropTypes.func.isRequired};
+import "../styles/Calculator.css"
 
 // Calculator Component
 const Calculator = () => {
